Add an `output_reference_idx` to `PhaseLinkingOptions`
This accounts for the fact that we may pass in several compressed SLCs for a
single run, and we wish to make the phase linking outputs referenced to
something besides the first SLC.

For example, if we had
   1_1_3
   1_4_6
   7_7_9
   10
   11
   12

And we wanted to output interferograms relative to `7`, we need to
specify that to the phase linking functions; otherwise, it will
output things relative to day 1. Since we dont write out rasters
corresponding to compressed SLC inputs, we'd have nothing from
day 7 to re-refrence later.

Note that an alternative might be to always write out rasters for
every input in a ministack, including the compressed inputs.
The downside is that during a normal local Sequential run with
multiple ministacks, you write out redundant rasters and need to
remove/select the right ones afterward.
But if that way seems cleaner later, we could revert this commit.