Refactor `compressed_reference_idx`, fix relative index bug (#453)
* Fix relative compressed reference index

Subtracting the `first_real_slc_idx` was not correct:
The stack planner would put 0 as the compressed index
for all ministacks.
Considering the 2nd ministack, if the `output_reference_idx`
was 0, then `compressed_reference_idx` would be 0. Then we
were slicing away the compressed, but also subtracting the
number of compressed. The led to *neagtive indexes*.

* add additional pre-set methods for making compressed SLCs

* shrink size of slcs to match in compress

* add comment, local test fix

* more comments