Remove single-burst special case, keep phase linking results to `phas… e487ce2  …e_linking/` (#574)
* Remove single-burst special case, keep phase linking results to `phase_linking/`

Currently the multi-burst outputs have one directory per burst ID, while the
single-burst is a special case that moves those contents to the top level
`work_directory`.
This is unnecessary, and has led to problems such as
https://github.com/opera-adt/disp-s1/issues/296

This always places the phase linking results in one directory up from
`work`, naming it `phase_linking` in case it's a general non-OPERA
swath.

* fix `test_workflows_displacement.py::test_separate_workflow_runs`